- Update samples for api changes so they compile and run - Update sample projects to suppress XamMac migration and reference the compiled MonoMac.dll instead of the one supplied by XS. - Mark a few methods in NSOpenGLContext as ThreadSafe, as used by the open gl samples

+++ samples/CFNetwork/Views/LogViewerController.cs
@@ -135,14 +135,14 @@ namespace MonoMac.CFNetwork.Test.Views {
 		{
 			var pos = Storage.Length;
 			Storage.Append (new NSAttributedString (text));
-			Text.SetFont (font, new NSRange (pos, text.Length));
+			Text.SetFont (font, new NSRange (pos, (uint)text.Length));
 		}
 
 		void DoAppend (NSColor color, string text)
 		{
 			var pos = Storage.Length;
 			Storage.Append (new NSAttributedString (text));
-			var range = new NSRange (pos, text.Length);
+			var range = new NSRange (pos, (uint)text.Length);
 			Text.SetFont (font, range);
 			Text.SetTextColor (color, range);
 		}
